Add community and Websockets

+++ application/account/routers/user_router.py
@@ -12,8 +12,7 @@ from application.account.crud import update_user, delete_user, upload_image_in_d
 from application.account.filters import UserFilter
 from application.account.helpers import get_current_active_user
 from application.account.models import User, IMAGE_DIR
-from application.account.schemas.user_schemas import UserReadSchema, UserUpdateSchema, UserReadSchemaShort, \
-    SubscriptionsSchema
+from application.account.schemas.user_schemas import UserReadSchema, UserUpdateSchema, UserReadSchemaShort
 from database.db import get_async_session
 
 disable_installed_extensions_check()
@@ -44,6 +43,9 @@ async def update_me(new_data: UserUpdateSchema, current_user: Annotated[User, De
                     session: AsyncSession = Depends(get_async_session)):
     new_data = new_data.model_dump(exclude_none=True)
 
+    if current_user.username == "test_user_ws":
+        raise HTTPException(status_code=status.HTTP_400_BAD_REQUEST, detail='You cannot update test user')
+
     await update_user(user=current_user, data=new_data, session=session)
 
     return await get_user(current_user.username)
@@ -52,6 +54,9 @@ async def update_me(new_data: UserUpdateSchema, current_user: Annotated[User, De
 @router.delete('/me', status_code=status.HTTP_204_NO_CONTENT)
 async def delete_me(current_user: Annotated[User, Depends(get_current_active_user)],
                     session: AsyncSession = Depends(get_async_session)):
+
+    if current_user.username == "test_user_ws":
+        raise HTTPException(status_code=status.HTTP_400_BAD_REQUEST, detail='You cannot delete test user')
 
     await delete_user(user=current_user, session=session)
 
@@ -121,6 +126,7 @@ async def get_account(account_id: int, current_user: Annotated[User, Depends(get
 @router.post('/subscribe/{account_id}', status_code=status.HTTP_202_ACCEPTED)
 async def subscribe(account_id: int, current_user: Annotated[User, Depends(get_current_active_user)],
                     session: AsyncSession = Depends(get_async_session)):
+
     if account_id not in current_user.subscriptions and account_id != current_user.id:
         current_user.subscriptions.append(account_id)
